Move *PackageFinder to the new 'discovery' module
Following up the discussion in #2887 and #2329, it seems that setuptools
is moving towards more automatic discovery features.

PackageFinder and PEP420PackageFinder are fundamental pieces of this
puzzle and grouping together them togheter with the code implementing these
new discovery features make a lot of sense.
